+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Blowing fuse".
 - Tests for "Container terminal".
 - Tests for "Fax machine".
+- Tests for "Jack Silver: the casino".
 
 ## [1.4.0] - 2022-03-03
 ### Added